Refactor layers to layers folder
