+++ CONTRIBUTING.md
@@ -4,12 +4,12 @@ _'murica!_
 
 The style is located within **src/** and is organized as follows:
 
-- **layer/** - Individual style layers, organized by subject area
+- **layer/** - Individual style layers, organized by subject area. The ordering of layers is specified in `index.js`.
 - **icons/** - SVG icons, which get converted into PNG stylesheets
 - **constants/** - Style elements that are frequently re-used
 - **js/** - Dynamic javascript code for highway shields and stylesheet building
-- **config.js** - Configuration settings (MapTiler API key, OpenMapTiles URL, etc)
-- **americana.js** - OpenMapTiles loader with layer ordering
+- **config.js** - Configuration settings (MapTiler API key, OpenMapTiles tile server URL, etc)
+- **americana.js** - MapLibre loader and configuration for the demo map
 - **index.html** - Demonstration map HTML page
 
 ## Install Node.js
